Fix Glissa's mana base
When playing last weekend, I had a really hard time getting the two
green for Glissa. Although there's about twice as many black cards
as green in the deck, most of the black has a high CMC and the deck
as a whole has tons of colorless cards, so I'm not too worried.